Add a "absolute" size mode to the image cleaner
On some deployments with DIND enabled the docker images from our builds
and the k8s dockerd are on the same partition. This means the GC of DIND
and k8s get in each others way. The "absolute" mode computes the size of
the docker directory and acts if it is too big in bytes.