Use buster as base image for image cleaner
Using Alpine as base distro means we have to build wheels for a lot of
packages which often requires a compiler and other native dependencies.
By using a debian base image we get wheels from pypi for free. The
BinderHub image uses buster already so while the individual image might
be larger the layers will be shared with an image that is already
present in the cluster.