Test updates, but still stuck at...
It seems that the image builds doesn't add to the space detected by the
image cleaner in our test setup.

If we don't run pytest as sudo, we get 0.0GB of space reported, and if
we use sudo, I get 0.1GB of space reported - never more. I've tried
using docker:20-dind and docker:24-dind without difference.

I think the key question is to figure out where in the dind container
the built images get saved.